+++ YahooFinanceApi.Tests/SummaryTests.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
 using Xunit;
 
 namespace YahooFinanceApi.Tests;
@@ -9,5 +10,8 @@ public class SummaryTests
     public async Task TestSummaryAsync()
     {
         var summary = await Yahoo.QuerySummaryAsync("MSFT", default, SummaryModules.AssetProfile, SummaryModules.DefaultKeyStatistics);
+        long? freeFloat = summary.GetValueOrDefault("FloatShares");
+        
+        Assert.NotNull(freeFloat);
     }
 }